authorization implemented

+++ src/api/notes/handler.js
@@ -7,8 +7,11 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this.validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this.service.addNote({ title, body, tags });
+    const noteId = await this.service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -17,12 +20,14 @@ class NotesHandler {
         noteId,
       },
     });
+    console.error(response);
     response.code(201);
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this.service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this.service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -33,6 +38,9 @@ class NotesHandler {
 
   async getNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this.service.verifyNoteOwner(id, credentialId);
     const note = await this.service.getNoteById(id);
     return {
       status: 'success',
@@ -46,7 +54,9 @@ class NotesHandler {
     this.validator.validateNotePayload(request.payload);
     const { title, body, tags } = request.payload;
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this.service.verifyNoteOwner(id, credentialId);
     await this.service.editNoteById(id, { title, body, tags });
 
     return {
@@ -57,6 +67,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this.service.verifyNoteOwner(id, credentialId);
     await this.service.deleteNoteById(id);
 
     return {
